feat(maintenance): tool to enable modern flows based on configuration (#39)
This allows for automated de-/activation of specific flows based on a given config. Additionally it allows for assigning the flows to a specified owner. For activation the option exists to use impersonation to factor in flow connection that are either not shared with the executing application user yet or an application user is not yet setup.